+++ src/Service/Pwa/Port/PwaService.mjs
@@ -73,15 +73,17 @@ export class PwaService {
 
     /**
      * @param {string} web_root
-     * @param {string} service_worker_file
+     * @param {string} service_worker_template_mjs_file
+     * @param {string} service_worker_mjs_file
      * @param {{[key: string]: *}} data
      * @returns {Promise<void>}
      */
-    async generateServiceWorker(web_root, service_worker_file, data) {
+    async generateServiceWorker(web_root, service_worker_template_mjs_file, service_worker_mjs_file, data) {
         await (await import("../Command/GenerateServiceWorkerCommand.mjs")).GenerateServiceWorkerCommand.new()
             .generateServiceWorker(
                 web_root,
-                service_worker_file,
+                service_worker_template_mjs_file,
+                service_worker_mjs_file,
                 data
             );
     }